+++ package.json
@@ -18,6 +18,7 @@
     "@types/node": "^18.13.0",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.10",
+    "@types/react-responsive": "^8.0.5",
     "@types/react-router-dom": "^5.3.3",
     "@types/swiper": "^6.0.0",
     "axios": "^1.3.4",
@@ -25,6 +26,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
+    "react-responsive": "^9.0.2",
     "react-router-dom": "^6.6.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
