+++ src/components/Nav/Mobile.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Mobile = () => {
+  return <div>mobile</div>;
+};
+
+export default Mobile;
