+++ src/components/Nav/Tablet.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Tablet = () => {
+  return <div>tablet</div>;
+};
+
+export default Tablet;
